+++ CHANGES.md
@@ -1,4 +1,6 @@
 
+- Add `Gg_kit.Color_scheme`. Note that some of the color schemes
+  are licensed under the Apache 2.0 license. 
 - Add `P2.orient_fast`.
 - Add `P2.seg_inter`.
 - Add `Float.seg_inter`.